py/text-from-image/pyproject.toml
[project]
name = "text-from-image"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "lolonao", email = "lolonao47@gmail.com" }
]
dependencies = [
    "tesseract>=0.1.3",
    "pillow>=10.4.0",
    "opencv-python>=4.10.0.84",
    "pytesseract>=0.3.13",
    "numpy>=2.1.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[tool.rye]
managed = true
virtual = true
dev-dependencies = []

[tool.pyright]
venvPath = "."
venv = ".venv"

